Refactored code format, improved consistency, re-coded solution in Utils files for dates for accuracy and efficiency, moved Chris_Integration methods into Utils as necessary
