+++ src/services/chris_integration.tsx
@@ -3,6 +3,7 @@ import ChrisAPIClient from "../api/chrisapiclient";
 import { ISeries, selectedImageType, StudyInstanceWithSeries } from "../context/reducers/analyseReducer";
 import { DcmImage } from "../context/reducers/dicomImagesReducer";
 import { PluginModels } from "../api/app.config";
+import { formatTime, modifyDatetime } from "../shared/utils"
 
 export interface LocalFile {
   name: string;
@@ -74,30 +75,6 @@ export const pollingBackend = async (pluginInstance: PluginInstance): Promise<Ba
   } else {
     return result;
   }
-}
-
-export const formatTime = (oldDay: string): string => {
-  return oldDay.split('.')[0]
-}
-
-export const modifyDatetime = (oldDay: string): string => {
-  let today = new Date().setHours(0, 0, 0, 0)
-  let diff = Math.abs(+today - +new Date(oldDay.split('T')[0]))
-  diff = Math.floor(diff / (1000 * 60 * 60 * 24)) // diff is in days, 1ms * 1000 * 60 * 60 * 24
-  let description = "days ago"
-  let rvtVal = `${diff} ${description}`
-  if (diff / 30 >= 1) {
-    description = diff / 30 >= 2 ? "months ago" : "month ago"
-    diff = Math.floor((diff / 30))
-    rvtVal = `${diff} ${description}`
-  } else if (diff / 7 >= 1) {
-    description = diff / 7 >= 2 ? "weeks ago" : "week ago"
-    diff = Math.floor(diff / 7)
-    rvtVal = `${diff} ${description}`
-  } else if (diff < 1) {
-    rvtVal = oldDay.split('T')[1].split('.')[0]
-  }
-  return rvtVal
 }
 
 // Dynamically loop through all model plug-ins and check if the current plug-in is valid
